Programmatically generate wordtable.h
src/wordtable.h contains a const char* definition of the word list as
strings separated by newline characters. In case any changes are made to
the source of the word list in plaintext/, there is a pair of python
scripts responsible for generating the header file based on the contents
of the word list.