Construct a map of word addresses
I wrote a helper program, construct_map, which takes the char array
already defined in src/wordtable.h, and finds the index within the array
of the first character of every word. This is to make it significantly
easier to access the nth word in the word table at runtime, as the 0th
address in the map is the 0th word, the 1st address is the 1st word,
etc.