Use 32-bit integer to store word addresses
I made the unaware mistake of using a 16-bit int to store an address in
a file over 350k large. I thought I was counting words (of which there
are less than 50 in the list), but I was in fact counting bytes. 20 bits
are more than enough to count past 1 million, but unfortunately
uint20_t, or let alone uint24_t is not a thing yet.