Use new comparison method for qsort
Use the first 2 characters of the string to make a comparison in qsort.
This is because the first 2 characters will be used to speed up the
search process when encoding words. Within the regions of words sorted
by the first 2 letters, the order of descending frequency will be
preserved. This will make it statistically faster to search for more
common words.