Create table of "guide words"
This is to speed up the process of searching for words, by indexing all
the words in the word list based on the first two characters in the
string that represents the word. Since each word is at least 2
characters long, including the null byte at the end, this works very
well.