Write function CMakeLists.txt files
Believe me when I tell you that this was a major pain in the behind. In
about a day I forced myself to understand, to the necessary extent, how
CMakeLists.txt and CMake as a whole work. Some stupid mistakes, like
forgetting to add include directories, slowed down my workflow
considerably. This is what happens when I have a 68 line Makefile that
I'm trying to convert into a CMakeLists.txt collection. It is not
pretty.

In other news, the rationale for this commit: moving toward a more
cross-platform build process. Makefiles had proven to be very difficult
to port to Windows on their own. Hopefully the time and frustration that
I have spent writing the CMakeLists.txt files will pay off.