Fix encoding repetition bug
Fixed the bug where when encoding words that existed in the word list,
the wordNum variable would only be initialized if the first character
group was made of word characters. The variable was not being written at
all in the main loop of the function, so the value of the first word (if
applicable) was used for all subsequent words.